line-personal-bot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pycryptodome==3.19.0
qrcode==7.4.2
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
starlette==0.27.0
typing-extensions==4.15.0
h11==0.16.0
anyio==3.7.1
sniffio==1.3.1
exceptiongroup==1.3.1
annotated-types==0.7.0
pydantic-core==2.14.1
pypng==0.20220715.0
charset-normalizer==3.4.7
